requirements.txt
npyscreen
packaging
pyyaml
